Move Types.kt into a new directory.
