+++ android/app/src/main/java/com/steve1316/uma_android_automation/bot/Trainee.kt
@@ -15,13 +15,13 @@ import com.steve1316.automation_library.utils.BotService
 import com.steve1316.uma_android_automation.MainActivity
 import com.steve1316.automation_library.utils.SettingsHelper
 import com.steve1316.uma_android_automation.utils.CustomImageUtils
-import com.steve1316.uma_android_automation.utils.types.StatName
-import com.steve1316.uma_android_automation.utils.types.Aptitude
-import com.steve1316.uma_android_automation.utils.types.RunningStyle
-import com.steve1316.uma_android_automation.utils.types.TrackSurface
-import com.steve1316.uma_android_automation.utils.types.TrackDistance
-import com.steve1316.uma_android_automation.utils.types.Mood
-import com.steve1316.uma_android_automation.utils.types.FanCountClass
+import com.steve1316.uma_android_automation.types.StatName
+import com.steve1316.uma_android_automation.types.Aptitude
+import com.steve1316.uma_android_automation.types.RunningStyle
+import com.steve1316.uma_android_automation.types.TrackSurface
+import com.steve1316.uma_android_automation.types.TrackDistance
+import com.steve1316.uma_android_automation.types.Mood
+import com.steve1316.uma_android_automation.types.FanCountClass
 import com.steve1316.uma_android_automation.components.ComponentInterface
 import com.steve1316.uma_android_automation.components.IconMoodGreat
 import com.steve1316.uma_android_automation.components.IconMoodGood
