+++ android/app/src/main/java/com/steve1316/uma_android_automation/bot/Training.kt
@@ -4,13 +4,13 @@ import android.util.Log
 import com.steve1316.uma_android_automation.MainActivity
 import com.steve1316.automation_library.utils.SettingsHelper
 import com.steve1316.uma_android_automation.utils.CustomImageUtils
-import com.steve1316.uma_android_automation.utils.types.StatName
-import com.steve1316.uma_android_automation.utils.types.Aptitude
-import com.steve1316.uma_android_automation.utils.types.RunningStyle
-import com.steve1316.uma_android_automation.utils.types.TrackSurface
-import com.steve1316.uma_android_automation.utils.types.TrackDistance
-import com.steve1316.uma_android_automation.utils.types.Mood
-import com.steve1316.uma_android_automation.utils.types.DateYear
+import com.steve1316.uma_android_automation.types.StatName
+import com.steve1316.uma_android_automation.types.Aptitude
+import com.steve1316.uma_android_automation.types.RunningStyle
+import com.steve1316.uma_android_automation.types.TrackSurface
+import com.steve1316.uma_android_automation.types.TrackDistance
+import com.steve1316.uma_android_automation.types.Mood
+import com.steve1316.uma_android_automation.types.DateYear
 import com.steve1316.uma_android_automation.utils.GameDate
 import com.steve1316.automation_library.data.SharedData
 import com.steve1316.automation_library.utils.BotService
