+++ android/app/src/main/java/com/steve1316/uma_android_automation/utils/CustomImageUtils.kt
@@ -13,10 +13,10 @@ import com.steve1316.automation_library.utils.SettingsHelper
 import com.steve1316.uma_android_automation.MainActivity
 import com.steve1316.uma_android_automation.bot.Game
 import com.steve1316.uma_android_automation.components.Region
-import com.steve1316.uma_android_automation.utils.types.StatName
-import com.steve1316.uma_android_automation.utils.types.Aptitude
-import com.steve1316.uma_android_automation.utils.types.BoundingBox
-import com.steve1316.uma_android_automation.utils.types.SkillData
+import com.steve1316.uma_android_automation.types.StatName
+import com.steve1316.uma_android_automation.types.Aptitude
+import com.steve1316.uma_android_automation.types.BoundingBox
+import com.steve1316.uma_android_automation.types.SkillData
 import com.steve1316.uma_android_automation.components.*
 import org.opencv.android.Utils
 import org.opencv.core.*
